+++ spring-test/src/main/java/org/springframework/test/context/CacheAwareContextLoaderDelegate.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2014 the original author or authors.
+ * Copyright 2002-2015 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,7 +23,7 @@ import org.springframework.test.annotation.DirtiesContext.HierarchyMode;
 /**
  * A {@code CacheAwareContextLoaderDelegate} is responsible for {@linkplain
  * #loadContext loading} and {@linkplain #closeContext closing} application
- * contexts, interacting transparently with a <em>context cache</em> behind
+ * contexts, interacting transparently with a {@link ContextCache} behind
  * the scenes.
  *
  * <p>Note: {@code CacheAwareContextLoaderDelegate} does not extend the
@@ -38,7 +38,7 @@ public interface CacheAwareContextLoaderDelegate {
 	 * Load the {@linkplain ApplicationContext application context} for the supplied
 	 * {@link MergedContextConfiguration} by delegating to the {@link ContextLoader}
 	 * configured in the given {@code MergedContextConfiguration}.
-	 * <p>If the context is present in the <em>context cache</em> it will simply
+	 * <p>If the context is present in the {@code ContextCache} it will simply
 	 * be returned; otherwise, it will be loaded, stored in the cache, and returned.
 	 * @param mergedContextConfiguration the merged context configuration to use
 	 * to load the application context; never {@code null}
@@ -50,7 +50,7 @@ public interface CacheAwareContextLoaderDelegate {
 
 	/**
 	 * Remove the {@linkplain ApplicationContext application context} for the
-	 * supplied {@link MergedContextConfiguration} from the <em>context cache</em>
+	 * supplied {@link MergedContextConfiguration} from the {@code ContextCache}
 	 * and {@linkplain ConfigurableApplicationContext#close() close} it if it is
 	 * an instance of {@link ConfigurableApplicationContext}.
 	 * <p>The semantics of the supplied {@code HierarchyMode} must be honored when
