+++ spring-test/src/test/java/org/springframework/test/context/ContextCacheTests.java
@@ -18,11 +18,11 @@ package org.springframework.test.context;
 
 import org.junit.Before;
 import org.junit.Test;
-
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.test.annotation.DirtiesContext.HierarchyMode;
 import org.springframework.test.context.support.AnnotationConfigContextLoader;
+import org.springframework.test.context.support.DefaultContextCache;
 import org.springframework.test.util.ReflectionTestUtils;
 
 import static org.junit.Assert.*;
@@ -39,7 +39,7 @@ import static org.springframework.test.context.ContextCacheTestUtils.*;
  */
 public class ContextCacheTests {
 
-	private ContextCache contextCache = new ContextCache();
+	private ContextCache contextCache = new DefaultContextCache();
 
 
 	@Before
